requirements.txt
# Core Dependencies
opencv-python==4.8.1.78
numpy==1.24.3
pyzmq==25.1.0
imutils==0.5.4
flask==2.3.2

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Performance & Monitoring
psutil==5.9.5
matplotlib==3.7.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0 